New object:  Session
Master module used to indicate when the session is ready.  all modules
should check ${Session.Ready} to see if it's safe to perform actions -
it introduces waits when you change systems, dock, undock, or warp.
Will most likely need to be renamed to not conflict with the built-in IS
Session object, but I wanted to commit this before I leave for work.